@raunaqmorarka 2024-05-28
<img width="1056" alt="Screenshot 2024-05-28 at 10 42 20 AM" src="https://github.com/trinodb/trino/assets/4344846/7786e538-b556-41ea-983d-ecea55ec373a">
~4.2% CPU time improvement in insert benchmarks